requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
snowflake-connector-python[secure-local-storage]>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.18.0
duckduckgo-search>=4.0.0